split functional components better
https://coveord.atlassian.net/browse/KIT-3196
