+++ packages/atomic/src/components/common/breadbox/breadcrumb-types.ts
@@ -1,0 +1,10 @@
+import {VNode} from '@stencil/core';
+
+export interface Breadcrumb {
+  facetId: string;
+  label: string;
+  formattedValue: string[];
+  state?: 'idle' | 'selected' | 'excluded';
+  content?: VNode;
+  deselect: () => void;
+}
